% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboard_primals.R
\name{add_line}
\alias{add_line}
\title{Adds a line to a whiteboard object}
\usage{
add_line(
  wb,
  name = NULL,
  fun = function(x) x,
  nstage = 1,
  domain = NULL,
  df = NULL,
  label = "right",
  labelopts = list(size = 18/ggplot2::.pt),
  ggopts = list(color = "#0066cc", size = 2)
)
}
\arguments{
\item{wb}{A whiteboard object.}

\item{name}{The name of the line to be used in labeling it on the whiteboard and referring to it when adding new elements.}

\item{nstage}{For animated whiteboards, which stage should be added to?}

\item{domain}{A numeric vector giving the full set of points on the x-axis to be evaluated in formulas. Defaults to the domain given for the existing board.}

\item{df}{If preferred, instead of \code{name, function, domain, range} you can provide a two-column \code{data.frame}, where the first column \code{x} gives the x-axis values, and the second column, given the name of the line, contains the y-axis values.}

\item{label}{Set to \code{''} to draw only the line by itself. Set to \code{'left'}, \code{'right'}, or \code{'both'} to put the name of the line near the first, last, or both points of the line.}

\item{labelopts}{A list of aesthetic values to be passed to \code{ggplot2::annotate} when adding the label. Note you can include a \code{label} option in here and use a different label than \code{name} on the line. Set font size by setting the \code{size} argument, dividing your desired font size by \code{ggplot2::.pt}.}

\item{ggopts}{List of aesthetic values to be passed on to \code{ggplot2::geom_line}, such as \code{color}, \code{size}, or \code{linetype}.}

\item{function}{A function that takes a single input and produces a single output, describing the line to be drawn.}
}
\description{
This function adds a line (not necessarily straight) to a whiteboard object.
}
