% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{axis_zero}
\alias{axis_zero}
\title{Whiteboard theme for fixed-axis graphs}
\usage{
axis_zero(
  x_axis = 0,
  y_axis = 0,
  quadrants = 1,
  x_min = -100,
  y_min = -100,
  x_max = 100,
  y_max = 100,
  x_title = "X",
  y_title = "Y",
  x_title_loc = c(x_max, -4 - 2 * (any(1:2 \%in\% quadrants) & any(3:4 \%in\%
    quadrants))),
  y_title_loc = c(-2 - 2 * (any(c(1, 4) \%in\% quadrants) & any(c(2, 3) \%in\%
    quadrants)), y_max),
  title_opts = list(),
  title_opts_x = NULL,
  title_opts_y = NULL,
  line_opts = list(),
  line_opts_x = NULL,
  line_opts_y = NULL,
  base_theme = ggplot2::theme_void,
  base_size = 18,
  base_family = "",
  base_line_size = base_size/10,
  base_rect_size = base_size/10
)
}
\arguments{
\item{x_axis}{The value of y at which the x-axis will be drawn.}

\item{y_axis}{The value of x at which the y-axis will be drawn.}

\item{x_min, x_max, y_min, y_max}{The edges of the graph on the x and y axes, for the purposes of figuring out how far to draw the axes and place the axis titles.}

\item{x_title_loc, y_title_loc}{The locations of the axis titles, in terms of actual x and y values.}

\item{title_opts, title_opts_x, title_opts_y}{A list of options to be passed to \code{ggplot2::annotate} for the axis titles, for both axes or only x or y specifically.}

\item{line_opts, line_opts_x, line_opts_y}{A list of options to be passed to \code{ggplot2::geom_segment} for the axis lines, for both axes or only x or y specifically.}

\item{base_theme}{The **ggplot2** theme on which to build the axis-zero graph. Note that if this is not \code{ggplot2::theme_void}, you will likely have to do a lot of adjustments in \code{ggplot2::theme()} afterwards to make it look good.}

\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{size for rect elements}
}
\description{
This is a ggplot2 theme designed for graphs in which axes are fixed at (0,0) or some other numbers. This is technically not a theme as it includes geometries as well. Use \code{theme()} to modify afterwards.
}
