% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_starters.R
\name{blank_board}
\alias{blank_board}
\title{Start a blank whiteboard}
\usage{
blank_board(
  domain = 0:100,
  range = NULL,
  theme = theme_board(),
  x_title = "X",
  y_title = "Y",
  title = NULL
)
}
\arguments{
\item{domain}{A numeric vector giving the full set of points on the x-axis to be evaluated in formulas.}

\item{range}{A two-element numeric vector giving the limits of the whiteboard's y-axis.}

\item{theme}{The **ggplot2** theme to be used. Set to \code{NULL} to use the default **ggplot2** theme. Other theme modifications can be added after the whiteboard is turned into a \code{ggplot} object with \code{print()} or \code{animate_whiteboard()}.}

\item{x_title}{Axis title for the x-axis.}

\item{y_title}{Axis title for the y-axis.}

\item{title}{Graph title.}
}
\description{
This function initializes a blank whiteboard object, which can then be added to.
}
