% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_intersect.R
\name{curve_intersect}
\alias{curve_intersect}
\title{Intersection Between Two Curves}
\usage{
curve_intersect(curve1, curve2, domain = NULL)
}
\arguments{
\item{curve1, curve2}{These are each either \code{data.frame}s where the first column contains the \code{x}-axis values of each line, and the second column contains the \code{y}-axis values, or functions with a single input and a single numerical output. If you mix a function with a data frame, the function will be evaluated at all of the \code{x}-axis points included in the \code{data.frame}.}

\item{domain}{If both \code{curve1} and \code{curve2} are functions, they will be evaluated over the domain of \code{domain}.}
}
\description{
This function finds a point that represents the intersection between two curves.
}
\details{
This code is lightly modified from the function written by Andrew Heiss [here](https://github.com/andrewheiss/supply-demand-ggplot).
}
