% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_print_static.R
\name{ggboard}
\alias{ggboard}
\title{Produce a static ggplot object from a whiteboard}
\usage{
ggboard(wb, nstage = 1)
}
\arguments{
\item{wb}{A whiteboard object.}

\item{nstage}{For animated whiteboards, which stage should be used? To animate all stages, use \code{board_animate()}.}
}
\description{
Takes a whiteboard object and produces a \code{ggplot} object that can be further customized.
}
