% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_print_static.R
\name{print.whiteboard}
\alias{print.whiteboard}
\title{Print a static whiteboard}
\usage{
\method{print}{whiteboard}(wb, nstage = 1)
}
\arguments{
\item{wb}{A whiteboard object.}

\item{nstage}{For animated whiteboards, which stage should be printed? To animate all stages, use \code{board_animate()}.}
}
\description{
Takes a whiteboard object and produces an image out of it. If you want a ggplot object that can be further customized, instead use \code{ggboard()}.
}
