% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_board}
\alias{theme_board}
\title{Default whiteboard theme}
\usage{
theme_board(
  base_size = 18,
  base_family = "",
  base_line_size = base_size/10,
  base_rect_size = base_size/10
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{Base}{size for rect elements}
}
\description{
This is a ggplot2 theme designed for use with ggwhiteboard. This is a complete theme that controls all non-data display. Use \code{theme()} to modify afterwards.
}
