% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fixed}
\alias{theme_fixed}
\title{Whiteboard theme for fixed-axis graphs}
\usage{
theme_fixed(
  x_axis = 0,
  y_axis = 0,
  base_size = 18,
  base_family = "",
  base_line_size = base_size/5,
  base_rect_size = base_size/5
)
}
\arguments{
\item{x_axis}{The value of y at which the x-axis will be drawn.}

\item{y_axis}{The value of x at which the y-axis will be drawn.}

\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{Base}{size for rect elements}
}
\description{
This is a ggplot2 theme designed for graphs in which axes are fixed at (0,0) or some other numbers. This is technically not a theme as it includes geometries as well. Use \code{theme()} to modify afterwards.
}
