% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_starters.R
\name{blank_board}
\alias{blank_board}
\title{Start a blank whiteboard}
\usage{
blank_board(
  domain = 0:100,
  range = NULL,
  axis_zero = TRUE,
  x_title = "X",
  y_title = "Y",
  title = NULL
)
}
\arguments{
\item{domain}{A numeric vector giving the full set of points on the x-axis to be evaluated in formulas.}

\item{range}{A two-element numeric vector giving the limits of the whiteboard's y-axis.}

\item{axis_zero}{Applies the **ggwhiteboard** pseudo-theme \code{axis_zero}, which forces the x and y axes to center at (0, 0), but also takes some thematic control from the user.}

\item{x_title}{Axis title for the x-axis.}

\item{y_title}{Axis title for the y-axis.}

\item{title}{Graph title.}
}
\description{
This function initializes a blank whiteboard object, which can then be added to.
}
